Refactor GUI modules into src package
